Improve BinaryEvaluatorSnippetTests(Engine)
- Extract AbstractSnippetTestsEngine from AbstractLanguageSnippetTestsEngine
  and reuse it for BinaryEvaluatorSnippetTestEngine.
- Rename BinaryEvaluatorSnippetTestEngine to BinaryEvaluatorSnippetTestsEngine.
- Rename expected output files to end in ".yml.pkl"
  to satisfy expectation of AbstractSnippetTestsEngine.
- Suppress bogus IntelliJ warnings in expected output files.
- Add trailing newline to expected output files.
