+++ pkl-server/src/test/files/SnippetTests/output/map.yml
@@ -1,6 +1,7 @@
+#file: noinspection YAMLIncompatibleTypes
 - 1
-- map
-- file:///$snippetsDir/input/map.pkl
+- map.yml
+- file:///$snippetsDir/input/map.yml.pkl
 - 
   - 
     - 16
