+++ pkl-server/src/test/files/SnippetTests/output/pair.yml
@@ -1,6 +1,7 @@
+#file: noinspection YAMLIncompatibleTypes
 - 1
-- pair
-- file:///$snippetsDir/input/pair.pkl
+- pair.yml
+- file:///$snippetsDir/input/pair.yml.pkl
 - 
   - 
     - 16
