+++ pkl-server/src/test/resources/META-INF/services/org.junit.platform.engine.TestEngine
@@ -1,1 +1,1 @@
-org.pkl.server.BinaryEvaluatorSnippetTestEngine
+org.pkl.server.BinaryEvaluatorSnippetTestsEngine
